+++ src/DAL/entities/Message.ts
@@ -1,7 +1,7 @@
 /* istanbul ignore file */
 import 'reflect-metadata';
 import { Entity, PrimaryGeneratedColumn, Column, Timestamp } from 'typeorm';
-import { SeverityLevels, LogComponent, AnalyticsMessageTypes } from './../../common/interfaces';
+import { SeverityLevels, LogComponent, AnalyticsMessageTypes } from '../../common/interfaces';
 
 @Entity()
 export class Message {
