+++ src/app.ts
@@ -1,4 +1,3 @@
-import 'reflect-metadata';
 import { Application } from 'express';
 import { DependencyContainer } from 'tsyringe';
 import { registerExternalValues, RegisterOptions } from './containerConfig';
